requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
langchain-anthropic==0.1.13
langchain-core==0.1.45
jinja2==3.1.2
python-dotenv==1.0.0
openai
python-docx
beautifulsoup4
html2text
pydantic_settings
httpx[http2]
weasyprint
pdfkit
